Revert "Redirects file updates"
