+++ plugins/gatsby-source-umich-lib/gatsby-node.js
@@ -2,10 +2,9 @@ const path = require(`path`)
 const { fetch } = require('./fetch')
 const { createBreadcrumb } = require(`./create-breadcrumb`)
 const { createStaffNodes } = require(`./create-staff-nodes`)
-const {
-  createNetlifyRedirectsFile,
-  createClientSideRedirects,
-} = require('./create-redirects')
+const https = require('https')
+const fs = require('fs')
+const readline = require('readline')
 
 /**
  * Implement Gatsby's Node APIs in this file.
@@ -204,13 +203,29 @@ exports.onCreateNode = async ({ node, actions }, { baseUrl }) => {
 // Implement the Gatsby API “createPages”. This is called once the
 // data layer is bootstrapped to let plugins create pages from data.
 exports.createPages = ({ actions, graphql }, { baseUrl }) => {
-  const { createPage } = actions
+  const readInterface = readline.createInterface({
+    input: fs.createReadStream('public/_redirects'),
+  })
+  const { createRedirect } = actions
+  readInterface.on('line', function(line) {
+    if (line) {
+      const urls = line.split(' ')
+      /*
+      console.log(
+        'Creating client-side redirect from ' + urls[0] + ' to ' + urls[1]
+      )
+      */
+      createRedirect({
+        fromPath: urls[0],
+        toPath: urls[1],
+        isPermanent: true,
+        redirectInBrowser: true,
+        force: true,
+      })
+    }
+  })
 
-  /**
-   * This is a useful backup if the server-side redirects ever fail.
-   * Redirects are very important!
-   */
-  createClientSideRedirects({ createRedirect: actions.createRedirect })
+  const { createPage } = actions
 
   return new Promise((resolve, reject) => {
     const basicTemplate = path.resolve(`src/templates/basic.js`)
@@ -631,6 +646,23 @@ exports.createPages = ({ actions, graphql }, { baseUrl }) => {
   })
 }
 
-exports.onPreBuild = async ({}, { baseUrl }) => {
-  createNetlifyRedirectsFile({ baseUrl: removeTrailingSlash(baseUrl) })
+exports.onPreBootstrap = async ({}, { baseUrl }) => {
+  const dir = 'public'
+  if (!fs.existsSync(dir)) {
+    fs.mkdirSync(dir)
+  }
+  const file = fs.createWriteStream('public/_redirects')
+  const url = removeTrailingSlash(baseUrl)
+
+  https.get(url + '/_redirects', function(response) {
+    response.pipe(file)
+    file.on('finish', function() {
+      console.log('_redirects file downloaded')
+    })
+    file.on('error', function() {
+      console.log('There was an error downloading _redirects file')
+    })
+  })
+
+  return
 }
